+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [2.19.0]
 - Moved delta scan preview setting to settings page.
 - New error message in UI when net new scan is done on an invalid repository. Net new scans only work on Git.
+- Clear in Memory cache when branch is changed.
+- Added Clear Persisted Cache command.
+- Add support for ai fix feedback analytic when pressing apply on a fix.
 
 ## [2.18.2]
 - Update Language Server Protocol version to 15.